+++ src-charts/plugin.js
@@ -1,8 +1,9 @@
 import chartHTML from 'inline:./embed/chart.html';
 import {FORMULA_CHART_CONFIG_DIALOG, TABLE_CHART_CONFIG_DIALOG} from "./constants.js";
 import {addWindowVariableToHtmlString} from "../common-utils/embed-helpers.js";
-import {cloneDeep, get} from "lodash-es";
+import {cloneDeep} from "lodash-es";
 import Formula from 'fparser';
+import {COMMON_EMBED_COMMANDS, createOnEmbedCallHandler} from "../common-utils/embed-comunication.js";
 
 const plugin = {
     insertText: {
@@ -39,7 +40,7 @@ const plugin = {
                     START_FROM_ZERO,
                     CHART_ASPECT_RATIO_SIZE
                 };
-                await this._validateChartConfig(chartData);
+                await plugin._validateChartConfig(chartData);
                 await app.context.replaceSelection(`<object data="plugin://${ app.context.pluginUUID }?${encodeURIComponent(JSON.stringify(chartData))}" data-aspect-ratio="${chartData.CHART_ASPECT_RATIO_SIZE}" />`);
                 return null;
             } catch (e) {
@@ -77,7 +78,7 @@ const plugin = {
                     START_FROM_ZERO: false,  // For future use
                     CHART_ASPECT_RATIO_SIZE
                 };
-                await this._validateChartConfig(chartData);
+                await plugin._validateChartConfig(chartData);
                 await app.context.replaceSelection(`<object data="plugin://${app.context.pluginUUID}?${encodeURIComponent(JSON.stringify(chartData))}" data-aspect-ratio="${chartData.CHART_ASPECT_RATIO_SIZE}" />`);
                 return null;
             } catch (e) {
@@ -95,120 +96,89 @@ const plugin = {
             return 'Error parsing object tag';
         }
     },
-    async onEmbedCall(app, commandName, ...args) {
-        console.log('onEmbedCall', commandName, args);
-        switch (commandName) {
-            case 'getSettings':
-                return { type: 'success', result: app.settings };
-            case 'getAppProp':
-                const propName = args[0];
-                return { type: 'success', result: get(app, propName) };
-            case 'saveFile':
-                try {
-                    let {name, data} = args[0];
-                    if (data.startsWith('data:')) { // if data is url, convert to blob
-                        const response = await fetch(data);
-                        data = await response.blob();
-                    }
-                    const saved = await app.saveFile(data, name);
-                    return { type: 'success', result: saved };
-                } catch (e) {
-                    return { type: 'error', result: e.message };
+    onEmbedCall : createOnEmbedCallHandler({
+        ...COMMON_EMBED_COMMANDS,
+        updateChartData: async (app, chartDataOld) => {
+            let chartDataNew = {};
+            if (chartDataOld.DATA_SOURCE === 'note') {
+                const dialog = cloneDeep(TABLE_CHART_CONFIG_DIALOG);
+                dialog[1].inputs[0].value = chartDataOld.CHART_TYPE;
+                dialog[1].inputs[1].value = chartDataOld.DATA_SOURCE_NOTE_UUID;
+                dialog[1].inputs[1].type = 'text';
+                dialog[1].inputs[2].value = chartDataOld.CHART_TITLE;
+                dialog[1].inputs[3].value = chartDataOld.TABLE_INDEX_IN_NOTE;
+                dialog[1].inputs[4].value = chartDataOld.HORIZONTAL_AXIS_LABEL_DIRECTION;
+                dialog[1].inputs[5].value = chartDataOld.START_FROM_ZERO;
+                dialog[1].inputs[6].value = chartDataOld.CHART_ASPECT_RATIO_SIZE;
+                const promptResult = await app.prompt(...dialog);
+                if (!promptResult) return;
+                let [
+                    CHART_TYPE,
+                    DATA_SOURCE_NOTE_UUID,
+                    CHART_TITLE,
+                    TABLE_INDEX_IN_NOTE,
+                    HORIZONTAL_AXIS_LABEL_DIRECTION,
+                    START_FROM_ZERO,
+                    CHART_ASPECT_RATIO_SIZE
+                ] = promptResult;
+                if (DATA_SOURCE_NOTE_UUID) {
+                    DATA_SOURCE_NOTE_UUID = await app.findNote({uuid: DATA_SOURCE_NOTE_UUID});
+                    DATA_SOURCE_NOTE_UUID = DATA_SOURCE_NOTE_UUID.uuid;
                 }
-            case 'updateChartObject':   // This is disgusting code - need to refactor
-                try {
-                    const chartDataOld = args[0];
-                    let chartDataNew = {};
-                    if (chartDataOld.DATA_SOURCE === 'note') {
-                        const dialog = cloneDeep(TABLE_CHART_CONFIG_DIALOG);
-                        dialog[1].inputs[0].value = chartDataOld.CHART_TYPE;
-                        dialog[1].inputs[1].value = chartDataOld.DATA_SOURCE_NOTE_UUID;
-                        dialog[1].inputs[1].type = 'text';
-                        dialog[1].inputs[2].value = chartDataOld.CHART_TITLE;
-                        dialog[1].inputs[3].value = chartDataOld.TABLE_INDEX_IN_NOTE;
-                        dialog[1].inputs[4].value = chartDataOld.HORIZONTAL_AXIS_LABEL_DIRECTION;
-                        dialog[1].inputs[5].value = chartDataOld.START_FROM_ZERO;
-                        dialog[1].inputs[6].value = chartDataOld.CHART_ASPECT_RATIO_SIZE;
-                        const promptResult = await app.prompt(...dialog);
-                        if (!promptResult) return;
-                        let [
-                            CHART_TYPE,
-                            DATA_SOURCE_NOTE_UUID,
-                            CHART_TITLE,
-                            TABLE_INDEX_IN_NOTE,
-                            HORIZONTAL_AXIS_LABEL_DIRECTION,
-                            START_FROM_ZERO,
-                            CHART_ASPECT_RATIO_SIZE
-                        ] = promptResult;
-                        if (DATA_SOURCE_NOTE_UUID) {
-                            DATA_SOURCE_NOTE_UUID = await app.findNote({uuid: DATA_SOURCE_NOTE_UUID});
-                            DATA_SOURCE_NOTE_UUID = DATA_SOURCE_NOTE_UUID.uuid;
-                        }
-                        chartDataNew = {
-                            ...chartDataOld,
-                            CHART_TYPE,
-                            DATA_SOURCE_NOTE_UUID,
-                            CHART_TITLE,
-                            TABLE_INDEX_IN_NOTE,
-                            HORIZONTAL_AXIS_LABEL_DIRECTION,
-                            START_FROM_ZERO,
-                            CHART_ASPECT_RATIO_SIZE
-                        };
-                    } else if (chartDataOld.DATA_SOURCE === 'formula') {
-                        const dialog = cloneDeep(FORMULA_CHART_CONFIG_DIALOG);
-                        dialog[1].inputs[0].value = chartDataOld.CHART_TYPE;
-                        dialog[1].inputs[1].value = chartDataOld.CHART_TITLE;
-                        dialog[1].inputs[2].value = chartDataOld.DATA_SOURCE_FORMULA_F;
-                        dialog[1].inputs[3].value = chartDataOld.MIN_X;
-                        dialog[1].inputs[4].value = chartDataOld.MAX_X;
-                        dialog[1].inputs[5].value = chartDataOld.STEP_X;
-                        dialog[1].inputs[6].value = chartDataOld.CHART_ASPECT_RATIO_SIZE;
-                        const promptResult = await app.prompt(...dialog);
-                        if (!promptResult) return;
-                        let [
-                            CHART_TYPE,
-                            CHART_TITLE,
-                            DATA_SOURCE_FORMULA_F,
-                            MIN_X,
-                            MAX_X,
-                            STEP_X,
-                            CHART_ASPECT_RATIO_SIZE
-                        ] = promptResult;
-                        chartDataNew = {
-                            ...chartDataOld,
-                            CHART_TYPE,
-                            CHART_TITLE,
-                            DATA_SOURCE_FORMULA_F,
-                            MIN_X,
-                            MAX_X,
-                            STEP_X,
-                            CHART_ASPECT_RATIO_SIZE
-                        };
-                    }
-                    await this._validateChartConfig(chartDataNew);
-
-                    const originalNoteContent = await app.getNoteContent({uuid: chartDataOld.RENDERING_NOTE_UUID});
-                    const originalEmbedTag = `<object data="plugin://${app.context.pluginUUID}?${encodeURIComponent(JSON.stringify(chartDataOld))}" data-aspect-ratio="${chartDataOld.CHART_ASPECT_RATIO_SIZE}" />`;
-                    if (!originalNoteContent.includes(originalEmbedTag)) {
-                        throw new Error('Original chart not found in note');
-                    }
-                    const newEmbedTag = `<object data="plugin://${app.context.pluginUUID}?${encodeURIComponent(JSON.stringify(chartDataNew))}" data-aspect-ratio="${chartDataNew.CHART_ASPECT_RATIO_SIZE}" />`;
-                    const newNoteContent = originalNoteContent.replace(originalEmbedTag, newEmbedTag);
-                    await app.replaceNoteContent({uuid: chartDataNew.RENDERING_NOTE_UUID}, newNoteContent);
-                    return { type: 'success', result: 'Chart updated' };
-                } catch (e) {
-                    console.log(e);
-                    app.alert('Error updating chart: ' + e.message);
-                }
-            default:
-                try {
-                    const result = await (_.get(app, commandName))(...args);
-                    return { type: 'success', result: result };
-                } catch (error) {
-                    return { type: 'error', result: error.message };
-                }
+                chartDataNew = {
+                    ...chartDataOld,
+                    CHART_TYPE,
+                    DATA_SOURCE_NOTE_UUID,
+                    CHART_TITLE,
+                    TABLE_INDEX_IN_NOTE,
+                    HORIZONTAL_AXIS_LABEL_DIRECTION,
+                    START_FROM_ZERO,
+                    CHART_ASPECT_RATIO_SIZE
+                };
+            } else if (chartDataOld.DATA_SOURCE === 'formula') {
+                const dialog = cloneDeep(FORMULA_CHART_CONFIG_DIALOG);
+                dialog[1].inputs[0].value = chartDataOld.CHART_TYPE;
+                dialog[1].inputs[1].value = chartDataOld.CHART_TITLE;
+                dialog[1].inputs[2].value = chartDataOld.DATA_SOURCE_FORMULA_F;
+                dialog[1].inputs[3].value = chartDataOld.MIN_X;
+                dialog[1].inputs[4].value = chartDataOld.MAX_X;
+                dialog[1].inputs[5].value = chartDataOld.STEP_X;
+                dialog[1].inputs[6].value = chartDataOld.CHART_ASPECT_RATIO_SIZE;
+                const promptResult = await app.prompt(...dialog);
+                if (!promptResult) return;
+                let [
+                    CHART_TYPE,
+                    CHART_TITLE,
+                    DATA_SOURCE_FORMULA_F,
+                    MIN_X,
+                    MAX_X,
+                    STEP_X,
+                    CHART_ASPECT_RATIO_SIZE
+                ] = promptResult;
+                chartDataNew = {
+                    ...chartDataOld,
+                    CHART_TYPE,
+                    CHART_TITLE,
+                    DATA_SOURCE_FORMULA_F,
+                    MIN_X,
+                    MAX_X,
+                    STEP_X,
+                    CHART_ASPECT_RATIO_SIZE
+                };
+            }
+            await plugin._validateChartConfig(chartDataNew);
+            console.log('ok', chartDataNew);
+            const originalNoteContent = await app.getNoteContent({uuid: chartDataOld.RENDERING_NOTE_UUID});
+            const originalEmbedTag = `<object data="plugin://${app.context.pluginUUID}?${encodeURIComponent(JSON.stringify(chartDataOld))}" data-aspect-ratio="${chartDataOld.CHART_ASPECT_RATIO_SIZE}" />`;
+            if (!originalNoteContent.includes(originalEmbedTag)) {
+                throw new Error('Original chart not found in note');
+            }
+            const newEmbedTag = `<object data="plugin://${app.context.pluginUUID}?${encodeURIComponent(JSON.stringify(chartDataNew))}" data-aspect-ratio="${chartDataNew.CHART_ASPECT_RATIO_SIZE}" />`;
+            const newNoteContent = originalNoteContent.replace(originalEmbedTag, newEmbedTag);
+            await app.replaceNoteContent({uuid: chartDataNew.RENDERING_NOTE_UUID}, newNoteContent);
+            return chartDataNew;
         }
-    },
+    }),
     async _validateChartConfig(chartData) {
         if (chartData.DATA_SOURCE !== 'note' && chartData.DATA_SOURCE !== 'formula') {
             throw new Error('Invalid data source');
