+++ src-mindmap/markmap/renderer.js
@@ -31,7 +31,7 @@ export async function initMarkMap() {
         markmap.setData(root);  // Needed to setData again to fix a bug with styling
     } catch (error) {
         console.error(error);
-        document.body.innerHTML = `<div style="color: red; font-size: 20px; padding: 20px;">${error.message}</div>`;
+        document.body.innerHTML = `<div style="color: red; font-size: 20px; padding: 20px;">Error: ${error.message}</div>`;
         throw error;
     }
 }
